Gracefully truncate long definitions at a safe length
The necessary function was added to Sopel in v6.6.2, so this plugin's
requirements can no longer simply accept any 6.x release.

In Sopel 7.1+, there is (or will be) a new feature that lets the bot
handle graceful truncation automatically via a new argument to `say()`.
Until it's time to drop older Sopel versions, this works well enough.