Fix regression for empty input
Intentional or not, it was a nice Easter egg before when doing just the
'.ud' command with no parameter would fetch and display the definition
for 'none' (currently, "None - [The lack] of [anything]").

My last patch (a way to request definitions other than the top one)
managed to break this. I didn't account for the empty-input case.
Consider this patch an apology; it restores the previous behavior.